% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{abcd_qest}
\alias{abcd_qest}
\title{Simulate streamflow using 'ABCD' Hydrology model (no snow)}
\usage{
abcd_qest(parm, P, PE, S_ini, G_ini, print.all = FALSE)
}
\arguments{
\item{parm}{abcd model parameters}

\item{P}{a vector of precipitation time-series (mm)}

\item{PE}{a vector of potential evaporation time-series (mm)}

\item{S_ini}{is the initial soil moisture (mm)}

\item{G_ini}{is the initial groundwater storage (mm)}
}
\value{
the output is a vector of PET values
}
\description{
\code{abcd_qest()} returns monthly or daily streamflow time-series
}

