% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{abcdCalibrate}
\alias{abcdCalibrate}
\title{Calibrate 'abcd' hydrology model}
\usage{
abcdCalibrate(..., Q.obs, metric = "KGE", na.rm = FALSE)
}
\arguments{
\item{...}{parameters to be passed to abcd_qest()}

\item{metric}{performance metric, either 'KGE', 'RMSE' or 'NSE'}

\item{na.rm}{logical parameter to remove NA values}
}
\value{
the performance measure of the abcd model
}
\description{
\code{abcd_calibrate()} returns calibration performance measures for abcd model
}

