% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcdQest.R
\name{abcdQest}
\alias{abcdQest}
\title{ABCD conceptual rainfall - runoff simulation model}
\usage{
abcdQest(parm, P, PE, S_ini, G_ini, print.all = FALSE)
}
\arguments{
\item{parm}{abcd model parameters}

\item{P}{a vector of precipitation time-series (mm)}

\item{PE}{a vector of potential evaporation time-series (mm)}

\item{S_ini}{is the initial soil moisture (mm)}

\item{G_ini}{is the initial groundwater storage (mm)}
}
\value{
the output is a time-series of run-off values
}
\description{
\code{abcd_qest()} returns monthly or daily streamflow time-series
 The ABCD water balance model is a simple hydrologic model for simulating
 streamflow in response to precipitation and potential evapotranspiration
 developed by Thomas (1981). The model is comprised of two storage
 compartmens: soil moisture and groundwater. The soil moisture gains water
 from precipitation and loses water to evapotranspiration (ET), surface
 runoff and groundwater recharge. The groundwater compartment gains water
 from recharge and loses water as discharge. The total streamflow is the sum
 of surface runoff from the soil moisture and groundwater discharge.
 The model has four parameters: a, b, c, a, d.
\itemize{
\item a ranges between 0 and 1.
\item b ranges between 260 - 1900 (Vandewiele et al. 1992).
\item c ranges between 0 - 1.
\item d ranges between 0 - 1.
}
}
