% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_econ.r
\name{ANNUALIZE}
\alias{ANNUALIZE}
\title{Annualize a time-sequence of costs}
\usage{
ANNUALIZE(data, r = 0.05)
}
\arguments{
\item{data}{placeholder}

\item{r}{discount rate}
}
\description{
Annualize a time-sequence of costs
}

