% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_econ.R
\name{ANNUALIZE_FROM_PV}
\alias{ANNUALIZE_FROM_PV}
\title{Annualized costs from present value}
\usage{
ANNUALIZE_FROM_PV(PV, r = 0.05, T_years)
}
\arguments{
\item{PV}{present value cost}

\item{r}{discount rate}

\item{T_years}{time horizon}
}
\description{
Annualized costs from present value
}

