% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_economics.R
\name{annualizedReturn}
\alias{annualizedReturn}
\title{Annualized costs from present value}
\usage{
annualizedReturn(data, r = 0.05)
}
\arguments{
\item{data}{present value term}

<<<<<<< HEAD
\item{r}{discount rate}
}
\value{
annualized return of investment
=======
\item{r}{discount rate (%)}
>>>>>>> b6090b6fc85e7c32fcf7dbf0d43a3caf405ea09c
}
\description{
Annualized costs from present value
}

