% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_wrappers.r
\name{boxcox_inverse}
\alias{boxcox_inverse}
\title{Inverse Box-cox transformation}
\usage{
boxcox_inverse(lambda, Y)
}
\arguments{
\item{Y}{box-cox transformation value}

\item{lamda}{value extracted from power transform}
}
\description{
\code{boxcox_inverse()} returns a sample based on knn scheme
}

