% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{boxcoxTranformInverse}
\alias{boxcoxTranformInverse}
\title{Inverse Box-cox transformation}
\usage{
boxcoxTranformInverse(lambda, Y)
}
\arguments{
\item{Y}{box-cox transformation value}

\item{lamda}{value extracted from power transform}
}
\description{
\code{boxcox_inverse()} returns a sample based on knn scheme
}

