% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wrappers.R
\name{colorBrewerAlpha}
\alias{colorBrewerAlpha}
\title{Add transperancy for the color brewer}
\usage{
colorBrewerAlpha(col, alpha = 1)
}
\arguments{
\item{col}{a vector of colors}

\item{alpha}{desired alpha level}
}
\value{
returns values
}
\description{
Add transperancy for the color brewer
}

