% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.R
\name{expand_grid}
\alias{expand_grid}
\title{Create a tibble from All Combinations of Factor Variables}
\usage{
expand_grid(...)
}
\arguments{
\item{...}{placeholder}
}
\description{
Create a tibble from All Combinations of Factor Variables
}

