% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.r
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Create a tibble from the Combinations of data frames}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{placeholder}
}
\description{
Create a tibble from the Combinations of data frames
}

