% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-munging.R
\name{expandGrid}
\alias{expandGrid}
\title{Create a tibble from All Combinations of Factor Variables}
\usage{
expandGrid(...)
}
\arguments{
\item{...}{vectors to be passed to the function}
}
\description{
Create a tibble from All Combinations of Factor Variables
}
