% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.r
\name{expandGridDf}
\alias{expandGridDf}
\title{Create a tibble from the Combinations of data frames}
\usage{
expandGridDf(...)
}
\arguments{
\item{...}{data frames to be passed to the function}
}
\description{
Create a tibble from the Combinations of data frames
}

