% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_wrappers.r
\name{extract_pval}
\alias{extract_pval}
\title{Function to extract the overall ANOVA p-value out of a linear model object}
\usage{
extract_pval(modelobject)
}
\arguments{
\item{modelobject}{is the model to extract p values from}
}
\description{
\code{extract_pval()} returns a sample based on knn scheme
}

