% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{extractPvalue}
\alias{extractPvalue}
\title{Extract ANOVA p-value out of a linear model object}
\usage{
extractPvalue(modelobject)
}
\arguments{
\item{modelobject}{is the model to extract p values from}
}
\description{
\code{extract_pval()} returns a sample based on knn scheme
}

