% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcmDataTransform.R
\name{gcmDataTransform}
\alias{gcmDataTransform}
\title{Transform raw GCM data to tidy format}
\usage{
gcmDataTransform(path = NULL, gridInfo = "cmip5grid.xlsx",
  climateVars = c("prcp", "tavg", "tmax", "tmin"),
  scenarios = c("historical", "rcp26", "rcp45", "rcp60", "rcp85"))
}
\arguments{
\item{path}{path to the GCM projections folder}

\item{gridInfo}{name of the grid info file containing gcm information}

\item{climateVars}{desired climate variables to be processed}

\item{scenarios}{climate scenarios to be included}
}
\description{
This script is tied to the outputs of the CMIP5 TOOL (Wi, 2017)
}
