% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.r
\name{GoodnessOfFit}
\alias{GoodnessOfFit}
\title{Calculate goodness of fit parameters}
\usage{
GoodnessOfFit(observed, simulated, output = "NSE")
}
\arguments{
\item{observed}{observed results}

\item{simulated}{simulation results}

\item{output}{measure to be calculated: NSE, KGE, RMSE or bias}
}
\description{
Calculate goodness of fit parameters
}

