% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.r
\name{grid_interpolate}
\alias{grid_interpolate}
\title{Interpolation from 2-D matrix}
\usage{
grid_interpolate(x, y, z = NULL, resolution = 100, ...)
}
\arguments{
\item{x}{parameter x}

\item{y}{parameter y}

\item{z}{parameter z}

\item{resolution}{placeholder}
}
\description{
Interpolation from 2-D matrix
}

