% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridInterpolate.R
\name{gridInterpolate}
\alias{gridInterpolate}
\title{Interpolation from 2-D matrix}
\usage{
gridInterpolate(x, y, z = NULL, resolution = 100, ...)
}
\arguments{
\item{x}{parameter on x-axis}

\item{y}{parameter on y-axis}

\item{z}{parameter on z-axis}

\item{resolution}{resolution used for interpolating z}
}
\description{
Interpolation from 2-D matrix
}
