% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro-modeling.R
\name{hamonPET}
\alias{hamonPET}
\title{Potential evapotranspiration (PET) by hamon method}
\usage{
hamonPET(tavg, Ld, KPEC)
}
\arguments{
\item{tavg}{average monthly temperature (DegC)}

\item{Ld}{length of day as multiple of 12 hours (for hamon equation)}

\item{KPEC}{calibration parameter (for hamon equation)}

\item{date}{a time-series date object}
}
\value{
returns monthly PET values (mm/month)
}
\description{
\url{http://data.snap.uaf.edu/data/Base/AK_2km/PET/Hamon_PET_equations.pdf}
}
