% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_wrappers.r
\name{knn}
\alias{knn}
\title{K nearest neighbors (KNN) bootrapping}
\usage{
knn(x, k, s)
}
\arguments{
\item{x}{is the prior value selected}

\item{k}{is the number of nearest points}

\item{s}{is the vector to sample from}
}
\description{
\code{knn()} returns a sample based on knn scheme
}

