% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wrappers.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Equation Display function for ggplot2}
\usage{
lm_eqn(df)
}
\arguments{
\item{df}{data frame to pass}
}
\value{
returns values
}
\description{
Equation Display function for ggplot2
}

