% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGOF.R
\name{modelGOF}
\alias{modelGOF}
\title{Calculate goodness of fit parameters}
\usage{
modelGOF(observed, simulated, output = "NSE")
}
\arguments{
\item{observed}{observed results}

\item{simulated}{simulation results}

\item{output}{measure to be calculated: NSE, KGE, RMSE or bias}
}
\description{
Calculate goodness of fit parameters
}
