% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.r
\name{multiple_replace}
\alias{multiple_replace}
\title{Replace multiple elements in a dataframe at once}
\usage{
multiple_replace(x, what, by)
}
\arguments{
\item{x}{placeholder}

\item{what}{placeholder}

\item{by}{placeholder}
}
\description{
Replace multiple elements in a dataframe at once
}

