% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-munging.R
\name{multiReplace}
\alias{multiReplace}
\title{Replace multiple elements in a dataframe at once}
\usage{
multiReplace(x, what, by)
}
\arguments{
\item{x}{placeholder}

\item{what}{placeholder}

\item{by}{placeholder}
}
\description{
Replace multiple elements in a dataframe at once
}
