% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{PET_calculate}
\alias{PET_calculate}
\title{Potential evapotranspiration (PET)}
\usage{
PET_calculate(date, tavg, tdif, lat, method = "hargreaves")
}
\arguments{
\item{date}{a time-series date object}

\item{tavg}{a vector of average temperature values (°C)}

\item{tdif}{a vector of differences computed from maximum and minimum
temperatures}

\item{lat}{latitude information (enter negative values for southern hemisphere)}

\item{method}{PET calculation method (currently only hargreaves formula is
used)}
}
\value{
the output is a vector of PET values
}
\description{
\code{PET_calculate()} returns monthly or daily PET values with the desired
calculation approach based on the input arguments
\url{https://en.wikipedia.org/wiki/Evapotranspiration}
}

