% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{PET_hamon}
\alias{PET_hamon}
\title{' Potential evapotranspiration (PET) by hamon method
\url{http://data.snap.uaf.edu/data/Base/AK_2km/PET/Hamon_PET_equations.pdf}}
\usage{
PET_hamon(date, tavg, Ld, KPEC)
}
\arguments{
\item{date}{a time-series date object}

\item{tavg}{average monthly temperature (DegC)}

\item{Ld}{length of day as multiple of 12 hours (for hamon equation)}

\item{KPEC}{calibration parameter (for hamon equation)}
}
\value{
returns monthly PET values (mm/month)
}
\description{
' Potential evapotranspiration (PET) by hamon method
\url{http://data.snap.uaf.edu/data/Base/AK_2km/PET/Hamon_PET_equations.pdf}
}

