% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-tools.R
\name{piecewiseLm}
\alias{piecewiseLm}
\title{Piecewise linear regression}
\usage{
piecewiseLm(var_name, knots)
}
\arguments{
\item{var_name}{variable name}

\item{knots}{position of knots}
}
\value{
RETURN_DESCRIPTION
}
\description{
Estimating continuous piecewise linear regression
\url{https://www.r-bloggers.com/estimating-continuous-piecewise-linear-regression/} #'
Assume you are given continuous predictor x and continuous predicted variable y. 
We want to estimate continuous piecewise linear regression with fixed knots stored 
in variable knots using standard lm procedure.
The key to a solution is proper definition of regression formula. In order to introduce #' possibility of change of slope in knot k we have to add a so called hinge term to the 
model max(0, x-k).
In the code given below function piece.formula automatically generates a proper right 
hand side of the regression formula given variable name and list of required knots. It #' is next tested on a simple function.
}
\examples{
# ADD_EXAMPLES_HERE
}
