% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_discovery.R
\name{prim_grid}
\alias{prim_grid}
\title{Generate a grid output of prim_boxes}
\usage{
prim_grid(input, y, threshold = 0, vdim = 3)
}
\arguments{
\item{input}{a matrix of input parameter values}

\item{y}{is a vector of response values}

\item{vdim}{is number of dimensions to explore in the input matrix}

\item{minimum}{risk threshold required for the prim boxes}
}
\description{
\code{prim_grid()} returns a data table of prim_boxes with given features of
mass, density, coverage, and weighted coverage
}

