% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim_wrappers.r
\name{prim_select}
\alias{prim_select}
\title{prim_select}
\usage{
prim_select(input, y, probs, var, box, ...)
}
\arguments{
\item{input}{a matrix of input parameter values}

\item{y}{is a vector of response values}

\item{probs}{placeholder}

\item{var}{placeholder}

\item{box}{placeholder}
}
\description{
\code{prim_grid()} returns a data table of prim_boxes with given features of
mass, density, coverage, and weighted coverage
}

