% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim_wrappers.R
\name{primGrid}
\alias{primGrid}
\title{Generate a grid output of prim_boxes}
\usage{
primGrid(input, y, probs, threshold = 0, vdim = 2, ...)
}
\arguments{
\item{input}{a matrix of input parameter values}

\item{y}{is a vector of response values}

\item{probs}{placeholder}

\item{vdim}{placeholder}

\item{...}{placeholder}
}
\description{
\code{primGrid()} returns results for many prim_boxes with given features of
mass, density, coverage, and weighted coverage
}

