% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_economics.r
\name{PV_CALCULATE}
\alias{PV_CALCULATE}
\title{Calculate the PV from annual cash flows}
\usage{
PV_CALCULATE(data, r = 0.05)
}
\arguments{
\item{data}{placeholder}

\item{r}{discount rate}
}
\description{
Calculate the PV from annual cash flows
}

