% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin_econ.R
\name{PV_FROM_ANNUALIZE}
\alias{PV_FROM_ANNUALIZE}
\title{present value crom annualized costs}
\usage{
PV_FROM_ANNUALIZE(ANNUALIZE, r = 0.05, T_years)
}
\arguments{
\item{r}{discount rate}

\item{T_years}{time horizon}

\item{PV}{present value cost}
}
\description{
present value crom annualized costs
}

