% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging.r
\name{rename_columns}
\alias{rename_columns}
\title{Rename data frame columns in a list of data frames}
\usage{
rename_columns(data_list, new_names)
}
\arguments{
\item{data_list}{placeholder}

\item{new_names}{placeholder}
}
\description{
Rename data frame columns in a list of data frames
}

