% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir-modeling.R
\name{reservoirSimulation}
\alias{reservoirSimulation}
\title{Reservoir simulation model}
\usage{
reservoirSimulation(beg.y = 2020, beg.m = 1, K, K_dead, pool_flood = 0,
  pool_cons = 0.7, pool_buff = 0.3, buffer = 1, Q, T_dom, T_irr, T_eco,
  evap.m, f_elev, f_vol, f_sarea, S_fr = 0.8, cycle = FALSE,
  C_alloc = 0.8, priority = c("T_eco", "T_dom", "T_irr"))
}
\arguments{
\item{beg.y}{First year of the simulation period [dimless]}

\item{beg.m}{First month of the simulation period [dimless]}

\item{K, }{Reservoir design storage capacity [MCM]}

\item{K_dead, }{top of inactive storage volume [MCM]}

\item{pool_flood}{relative size of flood protection pool}

\item{pool_cons}{relative size of conservation pool}

\item{pool_buff}{relative size of buffer pool}

\item{buffer}{reservoir storage buffer coefficient [dimless]}

\item{Q, }{Monthly inflows to reservoir [MCM/month]}

\item{T_dom, }{Domestic target release [MCM]}

\item{T_irr, }{Irrigation target release [MCM]}

\item{T_eco, }{Environmental target release [MCM]}

\item{evap.m}{per area evaporation from reservoir [MCM/month]}

\item{f_elev}{evalation as f(volume) [approxfun obj]}

\item{f_vol}{volume as f(elevation) [approxfun obj]}

\item{f_sarea}{s.area as f(volume) [approxfun obj]}

\item{S_fr}{initial fill ratio at the first month [dimless]}

\item{cycle}{time-series double cycling [T/F]}

\item{C_alloc}{allocation coefficient [dimless]}
}
\description{
\code{reservoir_sim()} returns simulation results from a reservoir analysis
}
