% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sceDefaults}
\alias{sceDefaults}
\title{Set sceDefaults}
\usage{
sceDefaults()
}
\arguments{
\item{ncomplex}{number of complexes}

\item{cce.iter}{number of iteration in inner loop (CCE algorithm)}

\item{fnscale}{function scaling factor (set to -1 for maximisation)}

\item{elitism}{controls amount of weighting in sampling towards the better parameter sets}

\item{initsample}{sampling scheme for initial values -- "latin" or "random"}

\item{reltol}{convergence threshold: relative improvement factor required in an SCE iteration}

\item{tolsteps}{number of iterations within reltol to confirm convergence}

\item{maxit}{maximum number of iterations}

\item{maxeval}{maximum number of function evaluations}

\item{maxtime}{maximum duration of optimization in seconds}

\item{returnpop}{whether to return populations from all iterations}

\item{trace}{level of user feedback}

\item{REPORT}{number of iterations between reports when trace >= 1}
}
\description{
Set sceDefaults
}

