% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{SCEoptim}
\alias{SCEoptim}
\title{Shuffled Complex Evolutionary (SCE) Algorithm}
\usage{
SCEoptim(FUN, par, ..., lower = -Inf, upper = Inf, control = list())
}
\arguments{
\item{FUN}{function to pass for optimization}

\item{par}{parameters to be optimized}

\item{lower}{lower bound for the parameters}

\item{upper}{upper bound for the parameters}

\item{control}{additional parameters}
}
\description{
Shuffled Complex Evolutionary (SCE) Algorithm
}

