% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought-indices.R
\name{waspIndex}
\alias{waspIndex}
\title{"Weighted Anomaly Standardized Precipitation" WASP index}
\usage{
waspIndex(data = NULL, N = NULL)
}
\arguments{
\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{N}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
WASP is an acronym for the "Weighted Anomaly Standardized 
Precipitation" index and is based solely on monthly precipitation data. 
It gives an estimate of the relative deficit or surplus of precipitation for 
different time intervals ranging from 1 to 12 months.
}
\details{
DETAILS
}
