% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletAnalysis.R
\name{waveletAnalysis}
\alias{waveletAnalysis}
\title{Wavelet Analysis Script (From Steinscheneider, 2013)}
\usage{
waveletAnalysis(variable, siglvl = 0.85, background_noise = "red")
}
\arguments{
\item{background_noise}{type of background noise Default:red}

\item{period}{Climate variable to be analyzed}

\item{sig}{sigicance level}
}
\value{
a list containing vectors period, signif_GWS, and GWS
}
\description{
Place holder
}
\details{
place holder
}
