% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin-economics.r
\name{annualizedReturn}
\alias{annualizedReturn}
\title{Annualized costs from present value}
\usage{
annualizedReturn(data, r = 0.05)
}
\arguments{
\item{data}{present value term}

\item{r}{discount rate}
}
\value{
annualized return of investment
}
\description{
Annualized costs from present value
}
