% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro-modeling.R
\name{GetwaterYear}
\alias{GetwaterYear}
\title{Water Year function}
\usage{
GetwaterYear(dates, start_month = 9)
}
\arguments{
\item{dates}{time-series of dates as a date object}

\item{start_month}{the first month of the water year}
}
\value{
the output is the calculated water year
}
\description{
\url{http://www.civil.uwaterloo.ca/watflood/manual/02_03_2.htm}
}
