% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro-modeling.R, R/stat-tools.R
\name{goodnessOfFit}
\alias{goodnessOfFit}
\alias{goodnessOfFit}
\title{Calculate goodness of fit parameters}
\usage{
goodnessOfFit(observed, simulated, output = "NSE")

goodnessOfFit(observed, simulated, output = "NSE")
}
\arguments{
\item{observed}{observed results}

\item{simulated}{simulation results}

\item{output}{measure to be calculated: NSE, KGE, RMSE or bias}

\item{observed}{observed results}

\item{simulated}{simulation results}

\item{output}{measure to be calculated: NSE, KGE, RMSE or bias}
}
\description{
Calculate goodness of fit parameters

Calculate goodness of fit parameters
}
