% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engin-economics.r
\name{presentValue}
\alias{presentValue}
\title{Calculate the PV from annual cash flows}
\usage{
presentValue(data, r = 0.05)
}
\arguments{
\item{data}{annual stream of benefits or costs}

\item{r}{discount rate}
}
\value{
returns prevesent value
}
\description{
Calculate the PV from annual cash flows
}
